updating api and specs
